+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.6.2
+    - Fixed preview georeference for Mapflow Imagery Search
+    - Fixed search results filtering after search provider change
+    - Search results are now cleared on exit from QGIS
+    - Force clearing of temporary files from previous launch on plugin startup
 ## 2.6.1
     - Fix python error on logout/login
 ## 2.6.0